btl_anim: give emitDrawEnvPackets a consistent byte-pointer prototype
emitDrawEnvPackets mixed a u32 *ot with a u8 *pkt, even though both are just
byte-address handles into the display-list buffer (ot is only ever handed to
addPrim, which reinterprets it as a P_TAG * internally). Type both parameters
and the return as u8 *, so the packet cursor and the OT pointer read uniformly
as byte pointers. Codegen-identical; SLUS_008.92 and all overlays still Match.

+++ src/btl_anim.c
@@ -5,7 +5,7 @@
 #include "battle.h"
 
 
-u8 *emitDrawEnvPackets(u32 *ot, u8 *pkt);
+u8 *emitDrawEnvPackets(u8 *ot, u8 *pkt);
 void callCdTick(void);
 void shutdownCardSubsystem(void);
 void initBattleSubsystems(void);
@@ -2020,7 +2020,7 @@ void copyDisplayCoords(DVECTOR *dst) {
  * @param pkt GPU packet allocation pointer.
  * @return Updated packet pointer (pkt + 24 bytes: two 12-byte packets).
  */
-u8* emitDrawEnvPackets(u32* ot, u8* pkt) {
+u8 *emitDrawEnvPackets(u8 *ot, u8 *pkt) {
     RECT rect;
     DR_AREA *area;
     DR_OFFSET *offset;
